ADD THE GAME() FUNCTION
This function will call the isUserWinner() functions which essentially executes a single round of the game,
and loops it five time. it also has a condition to replay a round if the previous round was a tie.
The function also keeps the scores of the two players and returns true if the user won and false if they lost.